Namespacing of JSON Exporter
